Fix macos compilation

+++ get_next_line/get_next_line_bonus.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   get_next_line_bonus.h                              :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: pepie <pepie@student.42.fr>                +#+  +:+       +#+        */
+/*   By: polepie <polepie@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/24 14:42:39 by pepie             #+#    #+#             */
-/*   Updated: 2023/12/15 13:41:02 by pepie            ###   ########.fr       */
+/*   Updated: 2024/02/05 21:25:22 by polepie          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -38,7 +38,7 @@ enum e_verif_nl {
 char	*get_next_line(int fd);
 
 void	ft_lstadd_back_create(t_list **lst, char *n);
-int		ft_lstiter_sum(void *lst, int (*f)(char *));
+int		ft_lstiter_sum(void *lst, unsigned long  (*f)(char *));
 char	*ft_strndup(char *src, int nb);
 char	*ft_freeman(t_gnl *gnl, char *ret);
 int		init_gnl(t_gnl *gnl);
